@@ -1,4 +1,4 @@
-// swift-tools-version:5.5
+// swift-tools-version:6.2.1
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
